platform/master/requirements.txt
django>=5.0
djangorestframework>=3.15
channels>=4.0
django-q2>=1.7
